Move the telephones management into donors
This commit completes the integration of Telephones into the main Donors source
code base (also renaming Donors to DonorRecords). This is due to its code
relation relevance. There is a lot of inter-dependency that justifies the
integration of them. The routes are adjusted too, with some additional RESTful
search methods (search by CPF and by name).
There are other data management code related to Vaccines, Travels, Addresses and
Cells that will be integrated on next commits.